services completos de pedidos

+++ pedidos/src/main/java/com/example/pedidos/models/Ciudad.java
@@ -21,7 +21,7 @@ public class Ciudad {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "latitud_ciudad")
     private double latitud;
